requirements.txt
attrs==19.3.0
environ-config==21.2.0
influxdb-client[ciso]==1.18.0
requests==2.24.0
